+++ django_afip/admin.py
@@ -5,9 +5,7 @@ from datetime import datetime
 from django.apps import apps
 from django.contrib import admin, messages
 from django.contrib.admin.sites import AlreadyRegistered
-from django.db.models import Count
 from django.http import HttpResponse
-from django.urls import reverse
 from django.utils.translation import ugettext as _
 
 from django_afip import exceptions, models
@@ -95,35 +93,11 @@ class ReceiptStatusFilter(admin.SimpleListFilter):
     def queryset(self, request, queryset):
         if self.value() == self.VALIDATED:
             return queryset.filter(
-                validation__result=models.Validation.RESULT_APPROVED
+                validation__result=models.ReceiptValidation.RESULT_APPROVED
             )
         if self.value() == self.NOT_VALIDATED:
             return queryset.exclude(
-                validation__result=models.Validation.RESULT_APPROVED
-            )
-
-
-class ReceiptBatchedFilter(admin.SimpleListFilter):
-    title = _('batched')
-    parameter_name = 'batched'
-
-    BATCHED = 'batched'
-    NOT_BATCHED = 'not_batched'
-
-    def lookups(self, request, model_admin):
-        return (
-            (self.BATCHED, _('Batched')),
-            (self.NOT_BATCHED, _('Not batched')),
-        )
-
-    def queryset(self, request, queryset):
-        if self.value() == self.BATCHED:
-            return queryset.filter(
-                batch__isnull=True,
-            )
-        if self.value() == self.NOT_BATCHED:
-            return queryset.filter(
-                batch__isnull=False
+                validation__result=models.ReceiptValidation.RESULT_APPROVED
             )
 
 
@@ -136,13 +110,10 @@ class ReceiptAdmin(admin.ModelAdmin):
         'receipt_number',
         'issued_date',
         'friendly_total_amount',
-        'batch_link',
         'validated',
     )
     list_filter = (
-        ReceiptBatchedFilter,
         ReceiptStatusFilter,
-        'batch',
     )
 
     __related_fields = [
@@ -171,7 +142,6 @@ class ReceiptAdmin(admin.ModelAdmin):
                 'validation',
                 'receipt_type',
                 'point_of_sales',
-                'batch',
             )
 
     def friendly_total_amount(self, obj):
@@ -182,7 +152,9 @@ class ReceiptAdmin(admin.ModelAdmin):
     friendly_total_amount.short_description = _('total amount')
 
     def validated(self, obj):
-        return obj.validation.result == models.Validation.RESULT_APPROVED
+        return (
+            obj.validation.result == models.ReceiptValidation.RESULT_APPROVED
+        )
     validated.short_description = _('validated')
     validated.admin_order_field = 'validation__result'
     validated.boolean = True
@@ -193,103 +165,14 @@ class ReceiptAdmin(admin.ModelAdmin):
     cae.admin_order_field = 'validation__cae'
 
     @catch_errors
-    def create_batch(self, request, queryset):
-        variations = queryset \
-            .aggregate(
-                receipt_types=Count('receipt_type', distinct=True),
-                points_of_sales=Count('point_of_sales', distinct=True),
-            )
-
-        if variations['receipt_types'] > 1:
-            self.message_user(
-                request,
-                _(
-                    'The selected receipts are not all of the same type.'
-                ),
-                messages.ERROR,
-            )
-            return
-        if variations['points_of_sales'] > 1:
-            self.message_user(
-                request,
-                _(
-                    'The selected receipts are not all from the same point '
-                    'of sales.'
-                ),
-                messages.ERROR,
-            )
-            return
-
-        queryset = queryset.filter(batch__isnull=True)
-        if not queryset.count():
-            self.message_user(
-                request,
-                _('All the selected invoices have already been batched'),
-                messages.ERROR,
-            )
-            return
-
-        models.ReceiptBatch.objects.create(queryset)
-        # TODO: Maybe redirect to batch screen?
-    create_batch.short_description = _('Create receipt batch')
-
-    def batch_link(self, obj):
-        if not obj.batch:
-            return None
-        return '<a href="{}">{}</a>'.format(
-            reverse('admin:afip_receiptbatch_change', args=(obj.batch.id,)),
-            obj.batch.id
-        )
-    batch_link.admin_order_field = 'batch'
-    batch_link.allow_tags = True
-    batch_link.short_description = _('batch')
-
-    actions = [create_batch]
-
-
-class ReceiptBatchAdmin(admin.ModelAdmin):
-    list_display = (
-        'id',
-        'receipts_count',
-        'validated',
-    )
-
-    def get_queryset(self, request):
-        queryset = super().get_queryset(request)
-        return queryset.annotate(
-            Count('receipts', distinct=True)
-        )
-
-    def validated(self, obj):
-        return obj.validation \
-            .filter(result=models.Validation.RESULT_APPROVED) \
-            .count() > 0
-    validated.short_description = _('validated')
-    validated.admin_order_field = 'validation__result'
-    validated.boolean = True
-
-    def receipts_count(self, obj):
-        return '<a href="{}?batch__id__exact={}">{}</a>'.format(
-            reverse(self.admin_site.name + ':afip_receipt_changelist'),
-            obj.id,
-            obj.receipts__count,
-        )
-    receipts_count.allow_tags = True
-    receipts_count.short_description = _('receipts')
-    receipts_count.admin_order_field = 'receipts__count'
-
-    @catch_errors
     def validate(self, request, queryset):
-        for batch in queryset:
-            errs = batch.validate()
-            if errs:
-                self.message_user(
-                    request,
-                    _(
-                        'Batch #%(num)s failed: %(err)s'
-                    ) % {'num': batch.pk, 'err': errs},
-                    messages.ERROR,
-                )
+        errs = queryset.validate()
+        if errs:
+            self.message_user(
+                request,
+                _('Receipt validation failed: %s') % errs,
+                messages.ERROR,
+            )
     validate.short_description = _('Validate')
 
     actions = [validate]